+++ backend/package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
+    "@json2csv/plainjs": "^7.0.6",
     "cors": "^2.8.5",
     "dotenv": "^17.2.3",
     "express": "^5.1.0",
